requirements.txt
aniso8601==9.0.1
APScheduler==3.8.1
argon2-cffi==21.1.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
cryptography==36.0.1
cymem==2.0.6
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fuzzywuzzy==0.18.0
google-search-results==2.4.0
greenlet==1.1.2
icd10-cm==0.0.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
msal==1.16.0
murmurhash==1.0.6
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pathy==0.6.1
preshed==3.0.6
#psycopg2==2.9.2
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
simple-icd-10==2.0.1
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.0.7
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.27
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.1.0
tqdm==4.62.3
typer==0.3.2
typing-extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.2
